==> chromosome_8/01_compare_arm_event_table.ipynb <==
# Make arm level event table for chromosome 8

Here we look for genes IN the event discussed that are affected by the arm-level event. We will find these effects by performing a series of t-tests comparing the proteomic values of the patients with the event against those without the event.

## Setup (Install necessary packages)

We will start by importing necessary packages and collecting all of the proteomics data we will need to run the tests. The cancer types analyzed should have been determined in 01_event_basic_info where we determine which types of cancer seem to have the event we are looking at.

import pandas as pd
import numpy as np
import os
import sys
import operator
import cptac
import cnvutils

cancer_types = [
    "brca",
    "colon",
    "hnscc",
    "lscc",
    "luad",
    "ovarian"
]

## Make the arm level event tables

news = {}

for cancer_type in cancer_types:
    
    # Get the event table
    event_file_path = os.path.join("..", "chromosome_7", "01_event_tables", f"{cancer_type}_cna_summary.tsv.gz")
    
    event = pd.\
    read_csv(event_file_path, sep='\t', dtype={"chromosome": str})
    
    event = event[event["chromosome"] == "8"].\
    melt(
        id_vars=["Patient_ID", "chromosome", "arm"]
    )
    
    event = event[
        (
            (event["arm"] == "p") & (event["variable"] == "prop_arm_deleted") |
            (event["arm"] == "q") & (event["variable"] == "prop_arm_amplified")
        )
    ].\
    drop(columns=["chromosome", "arm"]).\
    replace({
        "prop_arm_amplified": "gain_event",
        "prop_arm_deleted": "loss_event"
    })
    
    event = event.\
    assign(value=event["value"] >= 0.95).\
    pivot(
        index="Patient_ID",
        columns="variable",
        values="value"
    )
    
    event.columns.name = None
    event.index.name = None
    
    news[cancer_type] = event

## Load the previously generated event tables, which are based on crossover location

olds = {}

for cancer_type in cancer_types:
    
    df = pd.read_csv(f"{cancer_type}_has_event.tsv", sep="\t", index_col=0)
    olds[cancer_type] = df

## Compare the two

for cancer_type in cancer_types:
    
    old = olds[cancer_type]
    new = news[cancer_type]
    
    print(f"{cancer_type}: \n{(~new.eq(old)).sum()}\n")

for cancer_type in cancer_types:
    
    old = olds[cancer_type]
    new = news[cancer_type]
    
    print(f"""{cancer_type}
old gain: {old["gain_event"].sum()}
new gain: {new["gain_event"].sum()}

old loss: {old["loss_event"].sum()}
new loss: {new["loss_event"].sum()}
""")

for cancer_type in cancer_types:
    
    old = olds[cancer_type]
    new = news[cancer_type]

    print(f"""{cancer_type}
new - old gain: {new["gain_event"].sum() - old["gain_event"].sum()}
new - old loss: {new["loss_event"].sum() - old["loss_event"].sum()}
""")

# Make some lists
lists = {}

for cancer_type in cancer_types:
    
    old = olds[cancer_type]
    new = news[cancer_type]
    
    # loss only old
    loss_filter = old.loss_event.astype(float).combine(new.loss_event.astype(float), operator.sub)
    loo = loss_filter[loss_filter == 1].index
    
    # loss only new
    lon = loss_filter[loss_filter == -1].index    

    # gain only old
    gain_filter = old.gain_event.astype(float).combine(new.gain_event.astype(float), operator.sub)
    goo = gain_filter[gain_filter == 1].index
    
    # gain only new
    gon = gain_filter[gain_filter == -1].index
    
    # Save
    lists[cancer_type] = {}
    
    lists[cancer_type]["loo"] = loo
    lists[cancer_type]["lon"] = lon
    
    lists[cancer_type]["goo"] = goo
    lists[cancer_type]["gon"] = gon

for cancer_type in lists.keys():
    print(lists[cancer_type]["loo"])

for cancer_type in lists.keys():
    print(lists[cancer_type]["lon"])

for cancer_type in lists.keys():
    print(lists[cancer_type]["goo"])

for cancer_type in lists.keys():
    print(lists[cancer_type]["gon"])

dss = {
    "brca": cptac.Brca,
    "colon": cptac.Colon,
    "hnscc": cptac.Hnscc,
    "lscc": cptac.Lscc,
    "luad": cptac.Luad,
    "ovarian": cptac.Ovarian
}

def make_samples_plot(samples, chrm, cancer_type, datasets_dict):
    
    # Get CNV table and transpose
    ds = datasets_dict[cancer_type](no_internet=True)
    cnv = ds.get_CNV().transpose()
    
    # Join in the locations
    locs = cnvutils.\
    get_gene_locations().\
    dropna(axis="index", how="all")

    if isinstance(cnv.index, pd.MultiIndex):
        cnv = cnv.\
        join(locs, how="inner")
        
    else:
        cnv = cnv.\
        join(locs.droplevel(1), how="inner").\
        drop_duplicates()
        
    cnv = cnv.assign(loc=cnv[["start_bp", "end_cp"]].max(axis="columns"))

    # For each sample, select the genes on the chromosome we're interested in, and plot
    

for cancer_type in dss.keys():
    
    b = make_samples_plot(None, None, cancer_type, dss)

b.join(b[["start_bp"]].min(axis="columns").rename("min"))
